refactor: 🚚 DTOs a carpeta dto
- DTOs movidos a subcarpeta "dto"
- refactorizacion de los imports
